feat: status badge components
